asset view: Don't use a split view for the assets panel
This switches to simply using a box layout instead of a RutSplitView
that allows interactive dragging of the divider. Allowing interaction
here seems to make the UI more fussy rather than really being useful at
this point. The current idea anyway is that there should be a toggle
state instead for the asset view to temporarily claim a larger
proportion of the screen while navigating for assets.