split-view: don't allowing interactive dragging
The ability to interactively drag the divider for a split-view makes the
allocation process of the UI awkward considering cases where the user is
resizing the top-level window. There is an ambiguity over whether the
user's chosen divider position should be considered a fraction or a
pixel offset, and if an offset then from which direction.  So far the
ability to drag interactively hasn't proved useful so for now we are
simplifying the RutSplitView widget to only allow a fixed split fraction
instead.