Fix the position used for dragging entities
There were two problems which broke the position when dragging an
entity. Firstly is was using the wrong camera to work out the offset
vectors. It was taking the camera from the event which triggered the
drag. Presumably this would previously end up being the view camera
but at some point it changed to the global editor camera. This patch
just forces it to be the view camera.

Secondly if you begin the drag by selecting an entity then the
property inspector will start being displayed which will cause the
camera view to resize. The offset vectors were only being calculated
once with the original size so they would no longer be correct after
the inspector is displayed. This patch makes it maintain a pointer to
the translate grab closure from the RigCameraView and it will make
sure to update the offset vectors during allocate. This is a bit hacky
because it needs to flush the viewport on the view camera and update
the view transforms in order to get the right calculation.

Note: This highlights that it could be good to make the space for the
property inspector have a fixed size because it's quite annoying when
you're trying to drag and everything moves around.