Remove support for luminance textures
Luminance textures have been removed in GL3 so it is difficult to
support them on the GL3 driver unless we use the texture swizzle
extension like we do for alpha textures. However, it doesn't seem like
it would be worth the effort because we haven't yet hit a case in
practice where alpha textures wouldn't suffice in place of luminance
textures.

Reviewed-by: Robert Bragg <robert@linux.intel.com>