support selecting multiple entities
This replaces engine::selected_entity with engine::selected_entities and
operations that apply to the selection of entities are now applied to
all selected entities. A notable exception is that the manipulator tool
doesn't currently do anything smart when multiple entities are selected,
it simply relates to the last selected entity for now.