rig: add basic copy and paste support
This adds initial copy and paste support. RutShell now keeps track of a
current selection that must implement a new 'selectable' interface and
while there is something selected will handle ctrl-c/x key bindings to
be able to copy/cut the selection into the clipboard and from there
allow copying and pasting via ctrl-v. The actual pasting is implemented
with a 'mimable' interface and a new _DROP input event and the hope is
that we can also build on these for drag and drop support later.

Note: this doesn't currently integrate with the operating system's
native copy & paste apis so you can only copy and paste within Rig
itself.