Allow tracking multiple RigTransitions
Firstly this gives transitions a name property so we will be able to
distinguish them in the editor. Transitions are now ref-counted and in
the undo journal we no longer assume there is only one transition at
engine->selected_transition and instead operations are either applied to
a specific transition or all current transitions are taken into account.