Add color source material property
The interaction model of simply selecting an entity and then clicking an
asset and being able to automagically know how to apply the asset to the
entity is too limiting so now we are going to expose component
properties that can accept assets, starting with a "Color Source"
property for the material component.

This adds a new type of property so that the property::spec can specify
an asset type for validation and instrospection.

This adds an inspector for asset based properties which simply shows a
thumbnail of the asset. Currently there is no way to actually change
the asset via the inspector, but the plan is to make it possible to
select the thumbnail area so that when an asset is clicked, if an asset
inspector is selected then we will try and set the asset on the selected
property instead of applying it automatically to the entity.

Since this patch was starting to repeat a common pattern where we
wanted a widget that was simply a composition of other widgets this
adds a _COMPOSITE_SIZABLE interface and supporting utilities so making
composite widgets should be easier now.

Since this patch needed a way to show a thumbnail within a fixed sized
area this adds a new RutShim widget which can be given a fixed size and
have one child set on it whose preferred size will always be ignored
and it will just be allocated to the fixed size of the shim. Note: this
is similar to RutFixed except RutFixed never updates the size of
children. Potentially RutFixed could be changed to have no preferred
size instead of a fixed size since RutShim can be used to force a given
size instead.