journal: improve journal dump code
This changes the way a journal is dumped so we first dump the redo_list
followed by the reverse of the undo_list, showing a pointer between the
two. Visually when you look at this and press ctrl-z/y you can see the
pointer move up and down one list of operations. The operations can now
be indented so we can increase the indent when dumping a subjournal.

This also adds an assertion that we don't recursively insert into the
journal which would be an indication of a bug.