engine: add undo journal stack
This adds api for pushing/popping journals which will update
engine->undo_journal to the top of the stack. This makes it possible to
create sub journals and push them to catch operations that may be logged
by multiple orthogonal components and then log the sub journal as a
single operation.